campus_placement_prediction
This data set consists of Placement data of students in our campus. It includes secondary and higher secondary school percentage and specialization. It also includes degree specialization, type and Work experience and salary offers to the placed students
which group of student are going to placed and which one are no###
##label are status##

I have use Random forest, decision tree and logistic regression among of them the highest score of model is logistic regression with 84%